♻️ Refactor: text input 창 높이 관련 로직 hook으로 빼고 적용
