+++ src/components/Post/PostUpload.jsx
@@ -1,9 +1,10 @@
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import styled from 'styled-components';
 import defaultImg from '../../assets/images/basic-profile-m.png';
 import delteBtn from '../../assets/icons/delete.svg';
 import postImgUploadBtn from '../../assets/images/image-upload.png';
 import { getMultiImageSrcAPI } from '../../api/apis/image';
+import useHandleResizeHeight from '../../hooks/useHandleResizeHeight';
 
 const UserImageStyle = styled.img`
   width: 42px;
@@ -21,7 +22,7 @@ const PostWriteArticle = styled.article`
 const PostForm = styled.form`
   width: 100%;
   height: 100%;
-  padding-top: 12px;
+  padding: 12px 0;
 
   .post-input {
     display: block;
@@ -68,6 +69,7 @@ const ImageBox = styled.ul`
     }
   }
 `;
+
 const userImgSrc = () => {
   return localStorage.getItem('image');
 };
@@ -75,22 +77,18 @@ const userImgSrc = () => {
 function PostUpload({
   inputValue,
   setInputValue,
-  setDisabled,
+  setBtnDisabled,
   handleFormSubmit,
 }) {
+  const { ref, handleResizeHeight } = useHandleResizeHeight(inputValue);
   const [userprofile] = useState(userImgSrc);
-  const textarea = useRef();
-  const handleResizeHeight = () => {
-    textarea.current.style.height = 'auto';
-    textarea.current.style.height = textarea.current.scrollHeight + 'px';
-  };
 
   const handleValueChange = (e) => {
     const { id, value } = e.target;
     if (id !== 'image') {
       setInputValue({ ...inputValue, [id]: value });
-      if (value) setDisabled(false);
-      else setDisabled(true);
+      if (value) setBtnDisabled(false);
+      else setBtnDisabled(true);
     } else {
       const { files } = e.target;
       if (files.length + inputValue.image.length > 3)
@@ -120,12 +118,6 @@ function PostUpload({
     setInputValue({ ...inputValue, image: newImageList });
   };
 
-  useEffect(() => {
-    if (inputValue.content) {
-      handleResizeHeight();
-    }
-  }, [inputValue]);
-
   return (
     <>
       <UserImageStyle
@@ -143,7 +135,7 @@ function PostUpload({
             글 작성
           </label>
           <textarea
-            ref={textarea}
+            ref={ref}
             className="post-input"
             id="content"
             value={inputValue.content}
